+++ src/main/java/com/sprint5team/monew/base/util/NotificationTasklet.java
@@ -1,0 +1,22 @@
+package com.sprint5team.monew.base.util;
+
+import com.sprint5team.monew.domain.notification.service.NotificationService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.batch.core.StepContribution;
+import org.springframework.batch.core.scope.context.ChunkContext;
+import org.springframework.batch.core.step.tasklet.Tasklet;
+import org.springframework.batch.repeat.RepeatStatus;
+import org.springframework.stereotype.Component;
+
+@Component
+@RequiredArgsConstructor
+public class NotificationTasklet implements Tasklet {
+
+    private final NotificationService notificationService;
+
+    @Override
+    public RepeatStatus execute(StepContribution contribution, ChunkContext chunkContext) {
+        notificationService.notifyNewArticles();
+        return RepeatStatus.FINISHED;
+    }
+}
